+++ lib/water_drop/contracts/config.rb
@@ -3,7 +3,7 @@
 module WaterDrop
   module Contracts
     # Contract with validation rules for WaterDrop configuration details
-    class Config < Dry::Validation::Contract
+    class Config < Base
       # Ensure valid format of each seed broker so that rdkafka doesn't fail silently
       SEED_BROKER_FORMAT_REGEXP = %r{\A([^:/,]+:[0-9]+)(,[^:/,]+:[0-9]+)*\z}.freeze