+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # WaterDrop changelog
 
+## 2.1.1 (Unreleased)
+- Introduce a common base for validation contracts
+
 ## 2.1.0 (2022-01-03)
 - Ruby 3.1 support
 - Change the error notification key from `error.emitted` to `error.occurred`.